backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
dspy-ai==2.4.0
aiohttp==3.9.1
mcp==0.9.1
python-multipart==0.0.7
asyncio-mqtt==0.16.1
numpy==1.24.3
google-generativeai==0.3.2